+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * The search box has a new style.
 * Item moves and loadouts will now wait for each other, to prevent errors when they would collide. This means if you apply two loadouts, the second will wait for the first to complete before starting.
 * Item details are now toggled by clicking the "i" on the item popup, rather than just by hovering over it.
+* The Infusion Calculator has been removed, now that infusions are much more straightforward.
 
 # 3.4.1